requirements.txt
django>=3.2.0,<4.0.0
djangorestframework
Pillow
opencv-python
qrcode
pyzbar